Revert "West 55"
