+++ package.json
@@ -1,47 +1,15 @@
 {
-  "name": "ileadms",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@reach/router": "^1.3.4",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "axios": "^0.21.1",
-    "js-cookie": "^2.2.1",
-    "material-table": "^1.69.1",
-    "material-ui-search-bar": "^1.0.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.3.0",
-    "react-delay": "^0.1.0",
-    "react-dom": "^17.0.1",
-    "react-notification-timeline": "0.0.8",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "reactstrap": "^8.9.0"
-  },
+  "name": "westroad-admin",
+  "version": "1.0.0",
+  "description": "westroad admin app",
+  "main": "index.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+  "keywords": [
+    "westroad",
+    "amin"
+  ],
+  "author": "zivaka team",
+  "license": "UNLICENSED"
 }
