--- src/components/Test/Test1.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-const Test1 = () => {
-    return(
-        <>
-        <center>
-            <br />
-            <br />
-            <br />
-            <br />
-            <br />
-            <br />
-            <h1>Test 1</h1>
-        </center>
-        </>
-    );
-}
-
-export default Test1;
