--- src/components/Test/Test2.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-const Test2 = () => {
-    return(
-        <>
-        <center>
-        <br />
-            <br />
-            <br />
-            <br />
-            <br />
-            <br />
-            <h1>Test 2</h1>
-        </center>
-        </>
-    );
-}
-
-export default Test2;
