--- src/config/url.js
@@ -1,1 +1,0 @@
-export const BASE_URL="https://ilead.vaanic.com";
