--- src/index.css
@@ -1,10 +1,0 @@
-body {
- 
-    font-family: 'Roboto', sans-serif !important;
- 
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
